* master: (115 commits)
  Fixing ios images. Submitting v3.5.
  Fixing mongo btree munin graph.
  Adding Swift Reader to Goodies.
  Speeding up scrolling on feed list.
  Support navigating back to home page from action bar in Preferences
  Redrawing intelligence slider on ios.
  Normal profile.
  V3.0.4 (RC3)
  Massively speeding up the viewWillAppear on the feed list.
  RC2 (or is it RC3 by now?) of iOS app.
  Fixing a crash on mark read.
  Fixing broken story page control.
  Adding Pinterest.
  Tightening up comments line height.
  Fixing bug around loading feeds when there's no offline db.
  Correctly render DB changes in folder list more quickly.
  Fixing empty story hashes.
  Cleaning up kill utils.
  Fixing hiding and showing of notifier on fetch.
  iOS7 app RC2
  ...

+++ clients/ios/Classes/offline/OfflineSyncUnreads.m
@@ -66,29 +66,38 @@
                  ];
             }
         }
-        
     }];
     [appDelegate.database inTransaction:^(FMDatabase *db, BOOL *rollback) {
         // Once all unread hashes are in, only keep under preference for offline limit
-        NSInteger offlineLimit = [[NSUserDefaults standardUserDefaults] integerForKey:@"offline_store_limit"];
+        NSInteger offlineLimit = [[NSUserDefaults standardUserDefaults]
+                                  integerForKey:@"offline_store_limit"];
         NSString *order;
         NSString *orderComp;
-        if ([[[NSUserDefaults standardUserDefaults] objectForKey:@"default_order"] isEqualToString:@"oldest"]) {
+        if ([[[NSUserDefaults standardUserDefaults] objectForKey:@"default_order"]
+             isEqualToString:@"oldest"]) {
             order = @"ASC";
             orderComp = @">";
         } else {
             order = @"DESC";
             orderComp = @"<";
         }
-        FMResultSet *cursor = [db executeQuery:[NSString stringWithFormat:@"SELECT story_timestamp FROM unread_hashes ORDER BY story_timestamp %@ LIMIT 1 OFFSET %d", order, offlineLimit]];
+        NSString *lastStorySql = [NSString stringWithFormat:
+                                  @"SELECT story_timestamp FROM unread_hashes "
+                                  "ORDER BY story_timestamp %@ LIMIT 1 OFFSET %ld",
+                                  order, (long)offlineLimit];
+        FMResultSet *cursor = [db executeQuery:lastStorySql];
         int offlineLimitTimestamp = 0;
         while ([cursor next]) {
             offlineLimitTimestamp = [cursor intForColumn:@"story_timestamp"];
             break;
         }
-        NSLog(@"Deleting stories over limit: %d - %d", offlineLimit, offlineLimitTimestamp);
-        [db executeUpdate:[NSString stringWithFormat:@"DELETE FROM unread_hashes WHERE story_timestamp %@ %d", orderComp, offlineLimitTimestamp]];
-        [db executeUpdate:[NSString stringWithFormat:@"DELETE FROM stories WHERE story_timestamp %@ %d", orderComp, offlineLimitTimestamp]];
+        [cursor close];
+        
+        if (offlineLimitTimestamp) {
+            NSLog(@"Deleting stories over limit: %ld - %d", (long)offlineLimit, offlineLimitTimestamp);
+            [db executeUpdate:[NSString stringWithFormat:@"DELETE FROM unread_hashes WHERE story_timestamp %@ %d", orderComp, offlineLimitTimestamp]];
+            [db executeUpdate:[NSString stringWithFormat:@"DELETE FROM stories WHERE story_timestamp %@ %d", orderComp, offlineLimitTimestamp]];
+        }
     }];
     
     appDelegate.totalUnfetchedStoryCount = 0;
